Set up testing with jest and supertest for integration testing, added test folder, separated app logic from server starting logic
