+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "build": "npx tsc",
     "start": "node dist/index.js",
     "dev": "nodemon server.ts"
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@types/express": "^4.17.21",
+    "@types/supertest": "^6.0.2",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "nodemon": "^3.1.0",
