+++ server.ts
@@ -1,28 +1,16 @@
 require('dotenv').config();
-import express from 'express';
-import { Request, Response } from 'express';
-import { userRouter } from './routes/users';
-import { defaultErrorHandler, httpExceptionHandler, logErrors } from './middleware/error.middleware';
 import { initializeDB } from './services/db.services';
+import app from './app';
 
-const app = express();
+/*
 
-app.use(express.json()); // Parse incoming requests data
+    Here the server is started and db is initialized.
 
-app.use(logErrors); // middleware for logging errors
-app.use(httpExceptionHandler); // middleware for handling http errors
-app.use(defaultErrorHandler); // middleware for catching all other errors
-
-app.use("/users", userRouter); //endpoint specified
-
-app.get("/", (req: Request, res: Response) => {
-    res.status(200).send("Route good");
-})
-
+*/
 
 initializeDB()
-    .then(() => {
-        
+
+    .then(() => {    
         app.listen(3000, () => {
             console.log("server is running on 3000");
         })
@@ -32,4 +20,3 @@ initializeDB()
         process.exit(1);
     })
 
-export default app;
